FRANKBOT/requirements.txt.txt
python-binance
pandas
ta
requests